feature(U-6927237267): Prototype auth service
This commit provides the initialization of the Unity Foundation project. The main action includes creating Application.java as the entry point of the application, setting up the dependencies and plugins in build.gradle, and initializing the micronaut CLI in micronaut-cli.yml. The application is equipped with MySQL and Flyway for database management purposes. Besides, Gradle wrapper for building tasks is included for consistent building environment across systems.