feature(U-6932054538): Implement the /api/hasPermission endpoint
Updated the parameters on tenant and service entities to use the name instead of numerical IDs in the 'hasPermission' requests. Refactored tests accordingly in 'UnityIamTest.java'. Also, have added description field in the tenant structure in database schema and tests. Updated the authentication method to use executor service for blocking tasks.