Container image for migrator does not exist
Problem
-------

Deployments of the Unity Auth Service will need to migrate their
database. A convenient way of delivering this is a container image
that performs the necessary migrations. The Unity Auth Service does
not have a container image for migrating the database.

Solution
--------

Add a Dockerfile for a migrator image.